codeql fix

+++ pjmedia/src/test/codec_vectors.c
@@ -245,7 +245,7 @@ static int read_ITU_format(FILE  *fp_bitstream,
         /* increment j to skip over the number of bits in frame */
         j++;
 
-        for (i=0; i<number_of_16bit_words_per_frame; i++) {
+        for (i=0; i<(short)number_of_16bit_words_per_frame; i++) {
             packed_word = 0;
             bit_count = 15;
             while (bit_count >= 0) {
